The window has been made sizeable by moving the buttons for the current count result to the same row as the other buttons. This should provide a more robust and flexible GUI. MinHeight and MinWidth have been set to a size that will always show all input fields and all buttons.
An enumextension has been created to make using enum properties in the model possible.

The dropdownlists now use the related enums in the viewmodel. This makes these lists more robust because they will always contain the enum entries.

The textbox for the date has been replaced with a datepicker.

Column size of the metadata grid has been changed to allow the date of the count to be completely visable.