Various fixes and tweaks.
Created a base class for appstates in a way that it is possible to actions to happen before and after the action delegate. This way actions that have to happen for every state do not have be taken up in each state but rather is defined in the base class.
States that perform a custom action can now override the virtual 'On...' method. From the viewmodel we will call the non virtual method and pass an action delegate which in turn will call the virtual 'On...' method.